@sodonnel 2021-08-19
With this change the decommission tests ran:

```
[INFO] Running org.apache.hadoop.ozone.scm.node.TestDecommissionAndMaintenance
[INFO] Tests run: 11, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 520.23 s - in org.apache.hadoop.ozone.scm.node.TestDecommissionAndMaintenance
```

An older PR:

```
[INFO] Running org.apache.hadoop.ozone.scm.node.TestDecommissionAndMaintenance
[INFO] Tests run: 11, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 843.013 s - in org.apache.hadoop.ozone.scm.node.TestDecommissionAndMaintenance
```

So it looks like this saved 343 seconds (over 5 minutes). Also notable, the CI server runs these tests slower than my laptop locally.